requirements.txt
# AlphaGWAS Pipeline Requirements

# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
scipy>=1.9.0

# Genomics tools
pyliftover>=0.4      # Coordinate liftover
cyvcf2>=0.30.0       # Fast VCF parsing

# Data formats
pyarrow>=10.0.0      # Parquet support

# Optional: for API queries
requests>=2.28.0

# AlphaGenome (install separately per DeepMind instructions)
# alphagenome

# Development (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
